backend/functions/requirements.txt
firebase-functions==0.4.0
firebase-admin>=6.5.0
functions-framework>=3.5.0
fastapi
uvicorn
pandas
ta
yfinance
feedparser
requests
python-dotenv
google-generativeai
